Centralize rustc_driver management in `driver.rs`
